split header into its own file
